Speedup the Player which always selects the highest available node
We achieve this speedup by memorizing in the initialization of the player the
locations of all nodes, sorted in non-increasing order of the value of the node
at each location.

Then, when it's time to select the next move, we start from the last index we
selected, and continue walking the vector, until we find the next available
node.